improve c-version
