+++ schloesser/exercise05_06_07/bonus_c/.gitignore
