+++ schloesser/exercise05_06_07/bonus_c/Makefile
@@ -7,7 +7,7 @@ TARGET=main
 default: $(TARGET)
 all: default
 
-.PHONY: clean all default run $(TARGET)
+.PHONY: clean all default run $(TARGET) valgrind
 
 %.o: %.c $(HEADERS)
 	$(CC) $(CFLAGS) -c $< -o $@
@@ -17,6 +17,9 @@ $(TARGET):
 
 run: $(TARGET)
 	./main ../data/small.gph
+
+valgrind: $(TARGET)
+	valgrind ./main ../data/small.gph
 
 clean:
 	rm -f *.o main
