+++ schloesser/exercise05_06_07/bonus_c/runall.sh
@@ -9,6 +9,6 @@ for f in $FILES
 do
     echo "Processing file $f..."
     echo $f >> $OUTFILE 
-    echo "bonus c standalone" >> $OUTFILE
+    echo "bonus c" >> $OUTFILE
     (time ./main $f) &>> $OUTFILE
 done
